requirements.txt
Flask==2.0.1
Pillow==8.2.0
numpy==1.20.3
tensorflow==2.5.0
